# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test Backend Aksamedia",
        "description": "API Documentation for Technical Test Backend Aksamedia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://aksamedia-backed.laravel.cloud",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Masuk ke dalam aplikasi",
                "description": "Menukarkan username dan password dengan token Bearer.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|AbCdEf..."
                                                },
                                                "admin": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Gagal Autentikasi / Credential Salah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "unauthenticated"
                                        },
                                        "errors": {
                                            "properties": {
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "invalid credential"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Keluar dari aplikasi",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "logout success"
                                        },
                                        "data": {
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/divisions": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Ambil daftar divisi",
                "description": "Mengambil data divisi dengan fitur pagination dan pencarian.",
                "operationId": "getDivisions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman yang ingin ditampilkan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian nama divisi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "divisions": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DivisionResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "description": "Metadata Pagination",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login Request",
                "description": "Body request untuk login",
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "description": "Username pengguna",
                        "type": "string",
                        "example": "admin"
                    },
                    "password": {
                        "description": "Password pengguna",
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "Schema data user/admin",
                "properties": {
                    "id": {
                        "description": "ID User",
                        "type": "string",
                        "format": "uuid",
                        "example": "98a7s6d5-..."
                    },
                    "name": {
                        "description": "Nama Lengkap",
                        "type": "string",
                        "example": "Administrator"
                    },
                    "username": {
                        "description": "Username Login",
                        "type": "string",
                        "example": "admin"
                    },
                    "phone": {
                        "description": "Nomor Telepon",
                        "type": "string",
                        "example": "08123456789"
                    },
                    "email": {
                        "description": "Email User",
                        "type": "string",
                        "format": "email",
                        "example": "admin@aksamedia.com"
                    }
                },
                "type": "object"
            },
            "DivisionResource": {
                "title": "Division Resource",
                "description": "Schema data untuk object Divisi",
                "properties": {
                    "id": {
                        "description": "ID Divisi",
                        "type": "string",
                        "format": "uuid",
                        "example": "98a7s6d5-4f3g-2h1j-..."
                    },
                    "name": {
                        "description": "Nama Divisi",
                        "type": "string",
                        "example": "Mobile Development"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}